requirements.txt
easyocr
logging
numpy
pandas
matplotlib
scikit-learn
Pillow
opencv-python
wheel
albumentations